initial vite setup
