--- .cache/74/81aab0d5def3a0c702999a8f75f396.json
@@ -1,1 +1,0 @@
-{"id":"qU5I","dependencies":[{"name":"/home/liamegan/projects/wtc/preact-stickerbook/src/stickerbook.scss","includedInParent":true,"mtime":1679351095125},{"name":"/home/liamegan/projects/wtc/preact-stickerbook/package.json","includedInParent":true,"mtime":1679507068605}],"generated":{"css":".Stickerbook{overflow:hidden;position:relative}.Stickerbook__background,.Stickerbook__foreground,.Stickerbook__frame{height:100%;left:0;position:absolute;top:0;width:100%}.Stickerbook__background{background-position:50%;background-repeat:no-repeat;background-size:cover}.Stickerbook__foreground,.Stickerbook__frame{object-fit:cover}.Stickerbook__foreground{pointer-events:none}","js":""},"sourceMaps":{"css":{"mappings":[{"source":"stickerbook.scss","name":null,"original":{"line":1,"column":0},"generated":{"line":1,"column":0}},{"source":"stickerbook.scss","name":null,"original":{"line":2,"column":2},"generated":{"line":1,"column":13}},{"source":"stickerbook.scss","name":null,"original":{"line":2,"column":2},"generated":{"line":1,"column":28}},{"source":"stickerbook.scss","name":null,"original":{"line":3,"column":2},"generated":{"line":1,"column":29}},{"source":"stickerbook.scss","name":null,"original":{"line":6,"column":0},"generated":{"line":1,"column":47}},{"source":"stickerbook.scss","name":null,"original":{"line":9,"column":2},"generated":{"line":1,"column":117}},{"source":"stickerbook.scss","name":null,"original":{"line":9,"column":2},"generated":{"line":1,"column":128}},{"source":"stickerbook.scss","name":null,"original":{"line":10,"column":2},"generated":{"line":1,"column":129}},{"source":"stickerbook.scss","name":null,"original":{"line":10,"column":2},"generated":{"line":1,"column":135}},{"source":"stickerbook.scss","name":null,"original":{"line":11,"column":2},"generated":{"line":1,"column":136}},{"source":"stickerbook.scss","name":null,"original":{"line":11,"column":2},"generated":{"line":1,"column":153}},{"source":"stickerbook.scss","name":null,"original":{"line":12,"column":2},"generated":{"line":1,"column":154}},{"source":"stickerbook.scss","name":null,"original":{"line":12,"column":2},"generated":{"line":1,"column":159}},{"source":"stickerbook.scss","name":null,"original":{"line":13,"column":2},"generated":{"line":1,"column":160}},{"source":"stickerbook.scss","name":null,"original":{"line":16,"column":0},"generated":{"line":1,"column":171}},{"source":"stickerbook.scss","name":null,"original":{"line":17,"column":2},"generated":{"line":1,"column":196}},{"source":"stickerbook.scss","name":null,"original":{"line":17,"column":2},"generated":{"line":1,"column":219}},{"source":"stickerbook.scss","name":null,"original":{"line":18,"column":2},"generated":{"line":1,"column":220}},{"source":"stickerbook.scss","name":null,"original":{"line":18,"column":2},"generated":{"line":1,"column":247}},{"source":"stickerbook.scss","name":null,"original":{"line":19,"column":2},"generated":{"line":1,"column":248}},{"source":"stickerbook.scss","name":null,"original":{"line":22,"column":0},"generated":{"line":1,"column":270}},{"source":"stickerbook.scss","name":null,"original":{"line":24,"column":2},"generated":{"line":1,"column":315}},{"source":"stickerbook.scss","name":null,"original":{"line":27,"column":0},"generated":{"line":1,"column":332}},{"source":"stickerbook.scss","name":null,"original":{"line":28,"column":2},"generated":{"line":1,"column":357}}],"sources":{"stickerbook.scss":".Stickerbook {\n  overflow: hidden;\n  position: relative;\n}\n\n.Stickerbook__frame,\n.Stickerbook__foreground,\n.Stickerbook__background {\n  height: 100%;\n  left: 0;\n  position: absolute;\n  top: 0;\n  width: 100%;\n}\n\n.Stickerbook__background {\n  background-position: center;\n  background-repeat: no-repeat;\n  background-size: cover;\n}\n\n.Stickerbook__frame,\n.Stickerbook__foreground {\n  object-fit: cover;\n}\n\n.Stickerbook__foreground {\n  pointer-events: none;\n}\n"},"lineCount":null}},"error":null,"hash":"a906869958d5a6e7abe5e429335e8c9c","cacheData":{"env":{}}}
