--- .cache/96/d3191e22529965b2e262d68221467d.json
@@ -1,1 +1,0 @@
-{"id":"E8iT","dependencies":[{"name":"/home/liamegan/projects/wtc/preact-stickerbook/package.json","includedInParent":true,"mtime":1679507403685},{"name":"/home/liamegan/projects/wtc/preact-stickerbook/node_modules/wtc-simplegl/package.json","includedInParent":true,"mtime":1679440154449}],"generated":{"js":"\"use strict\";Object.defineProperty(exports,\"__esModule\",{value:!0}),exports.createBuffer=void 0;const e=(e,r,t=e.STATIC_DRAW,f=e.ARRAY_BUFFER)=>{const u=e.createBuffer();return e.bindBuffer(f,u),e.bufferData(f,r,t),u};exports.createBuffer=e;"},"sourceMaps":{"js":{"mappings":[{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","original":{"line":11,"column":2},"generated":{"line":1,"column":0}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","original":{"line":11,"column":2},"generated":{"line":1,"column":13}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","original":{"line":11,"column":2},"generated":{"line":1,"column":20}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","original":{"line":11,"column":2},"generated":{"line":1,"column":35}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","original":{"line":11,"column":2},"generated":{"line":1,"column":43}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","original":{"line":11,"column":2},"generated":{"line":1,"column":56}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","original":{"line":11,"column":2},"generated":{"line":1,"column":57}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","original":{"line":11,"column":2},"generated":{"line":1,"column":64}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","original":{"line":11,"column":2},"generated":{"line":1,"column":68}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","original":{"line":11,"column":2},"generated":{"line":1,"column":76}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","original":{"line":11,"column":2},"generated":{"line":1,"column":94}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","original":{"line":1,"column":0},"generated":{"line":1,"column":96}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"createBuffer","original":{"line":1,"column":6},"generated":{"line":1,"column":102}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"createBuffer","original":{"line":1,"column":21},"generated":{"line":1,"column":104}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"gl","original":{"line":2,"column":2},"generated":{"line":1,"column":105}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"data","original":{"line":3,"column":2},"generated":{"line":1,"column":107}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"usage","original":{"line":4,"column":2},"generated":{"line":1,"column":109}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"gl","original":{"line":4,"column":10},"generated":{"line":1,"column":111}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"STATIC_DRAW","original":{"line":4,"column":13},"generated":{"line":1,"column":113}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"type","original":{"line":5,"column":2},"generated":{"line":1,"column":125}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"gl","original":{"line":5,"column":9},"generated":{"line":1,"column":127}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"ARRAY_BUFFER","original":{"line":5,"column":12},"generated":{"line":1,"column":129}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","original":{"line":7,"column":2},"generated":{"line":1,"column":145}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"buffer","original":{"line":7,"column":8},"generated":{"line":1,"column":151}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"gl","original":{"line":7,"column":17},"generated":{"line":1,"column":153}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"createBuffer","original":{"line":7,"column":20},"generated":{"line":1,"column":155}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","original":{"line":10,"column":2},"generated":{"line":1,"column":170}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"gl","original":{"line":8,"column":2},"generated":{"line":1,"column":177}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"bindBuffer","original":{"line":8,"column":5},"generated":{"line":1,"column":179}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"type","original":{"line":8,"column":16},"generated":{"line":1,"column":190}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"buffer","original":{"line":8,"column":22},"generated":{"line":1,"column":192}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"gl","original":{"line":9,"column":2},"generated":{"line":1,"column":195}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"bufferData","original":{"line":9,"column":5},"generated":{"line":1,"column":197}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"type","original":{"line":9,"column":16},"generated":{"line":1,"column":208}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"data","original":{"line":9,"column":22},"generated":{"line":1,"column":210}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"usage","original":{"line":9,"column":28},"generated":{"line":1,"column":212}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"buffer","original":{"line":10,"column":9},"generated":{"line":1,"column":215}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"exports","original":{"line":11,"column":2},"generated":{"line":1,"column":218}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"createBuffer","original":{"line":11,"column":2},"generated":{"line":1,"column":226}},{"source":"../node_modules/wtc-simplegl/lib/createBuffer.js","name":"createBuffer","original":{"line":11,"column":2},"generated":{"line":1,"column":239}}],"sources":{"../node_modules/wtc-simplegl/lib/createBuffer.js":"const createBuffer = (\n  gl,\n  data,\n  usage = gl.STATIC_DRAW,\n  type = gl.ARRAY_BUFFER\n) => {\n  const buffer = gl.createBuffer();\n  gl.bindBuffer(type, buffer);\n  gl.bufferData(type, data, usage);\n  return buffer;\n};\n\nexport { createBuffer };\n"},"lineCount":null}},"error":null,"hash":"6baa4741f5020f21917860236f8c294b","cacheData":{"env":{}}}
