--- babel.config.json
@@ -1,10 +1,0 @@
-{
-  "plugins": [
-    [
-      "@babel/plugin-transform-react-jsx",
-      {
-        "pragma": "h"
-      }
-    ]
-  ]
-}
