--- old.babel
@@ -1,1 +1,0 @@
-{ "plugins": [["transform-react-jsx", { "pragma": "h" }]] }
