--- package copy.json
@@ -1,50 +1,0 @@
-{
-  "name": "@wethegit/preact-stickerbook",
-  "version": "1.0.4",
-  "description": "Easily create collage apps that are accessible by default.",
-  "source": "src/index.js",
-  "main": "dist/preact-stickerbook.js",
-  "umd:main": "dist/preact-stickerbook.umd.js",
-  "unpkg": "dist/preact-stickerbook.umd.js",
-  "module": "dist/preact-stickerbook.module.js",
-  "exports": {
-    ".": "./dist/preact-stickerbook.modern.js",
-    "./helpers": "./dist/preact-stickerbook-helpers.modern.js"
-  },
-  "style": "dist/preact-stickerbook.css",
-  "css": "dist/preact-stickerbook.css",
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "rm -rf dist && microbundle build && microbundle build -i src/helpers.js -o dist/preact-stickerbook-helpers.js && microbundle build -i src/use-stickerbook.js -o dist/preact-use-stickerbook.js",
-    "prepare": "npm run build"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/wethegit/preact-stickerbook.git"
-  },
-  "keywords": [
-    "preact",
-    "preact-component"
-  ],
-  "publishConfig": {
-    "access": "public"
-  },
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/wethegit/preact-stickerbook/issues"
-  },
-  "homepage": "https://github.com/wethegit/preact-stickerbook#readme",
-  "peerDependencies": {
-    "preact": "^10.5.12"
-  },
-  "devDependencies": {
-    "microbundle": "^0.13.0",
-    "sass": "^1.32.7"
-  },
-  "dependencies": {
-    "wtc-math": "^0.0.4"
-  }
-}
