+++ src/README.md
@@ -1,0 +1,3 @@
+## Playground
+
+Files in the root of this directory are not shipped as part of the library, use it as development environment to run and test the package.
