--- src/index.js
@@ -1,26 +1,0 @@
-import { h } from "preact";
-import {
-  classnames,
-  loadUrlAsImage,
-  coverCanvas,
-  exportStickerbook,
-  reorderSticker,
-  addSticker,
-  deleteSticker,
-  patchSticker,
-} from "./helpers";
-import { renderSticker } from "./renderSticker";
-// import { Stickerbook } from "./stickerbook";
-import { Sticker, Stickerbook } from "./sticker";
-
-export {
-  classnames,
-  loadUrlAsImage,
-  coverCanvas,
-  exportStickerbook,
-  reorderSticker,
-  addSticker,
-  deleteSticker,
-  patchSticker,
-  renderSticker, Sticker, Stickerbook
-};
