--- src/index.scss
@@ -1,2 +1,0 @@
-@import url(./sticker.scss);
-@import url(./stickerbook.scss);
