+++ src/lib/helpers/addSticker.js
@@ -1,0 +1,10 @@
+export function addSticker(stickers = [], sticker) {
+  if (!sticker) throw Error("No `sticker` provided");
+
+  // we add a unique key based on the time because otherwise the
+  // sticker will loose its state when re-rendering it, even though
+  // we save the state globally, this is a good failsafe
+  return stickers.concat([
+    { ...sticker, order: stickers.length, key: Date.now() },
+  ]);
+}
