+++ src/lib/helpers/classnames.js
@@ -1,0 +1,6 @@
+export function classnames(namesArray) {
+  return namesArray
+    .filter((v) => v != "")
+    .join(" ")
+    .trim();
+}
