+++ src/lib/helpers/index.js
@@ -1,0 +1,7 @@
+export { addSticker } from './addSticker'
+export { coverCanvas } from './coverCanvas'
+export { deleteSticker } from './deleteSticker'
+export { exportStickerbook } from './exportStickerbook'
+export { loadUrlAsImage } from './loadUrlAsImage'
+export { patchSticker } from './patchSticker'
+export { reorderSticker } from './reorderSticker'
