+++ src/lib/helpers/loadUrlAsImage.js
@@ -1,0 +1,10 @@
+// Helpers function to create images from paths
+export function loadUrlAsImage(item) {
+  return new Promise((resolve, reject) => {
+    const img = new Image();
+    img.crossOrigin = "anonymous";
+    img.onload = () => resolve(img);
+    img.onerror = (err) => reject(err);
+    img.src = item;
+  });
+}
