+++ src/lib/index.js
@@ -1,0 +1,2 @@
+export { default as Stickerbook } from "./stickerbook";
+export { default as Sticker } from "./sticker";
