+++ src/main.jsx
@@ -1,0 +1,4 @@
+import { render } from 'preact'
+import { App } from './app.jsx'
+
+render(<App />, document.getElementById('app'))
